Refactor series types to own file
